+++ Barnamenevisan.Localizing/Repository/LocalizedPropertyRepository.cs
@@ -3,7 +3,7 @@ using Microsoft.EntityFrameworkCore;
 
 namespace Barnamenevisan.Localizing.Repository
 {
-    public class LocalizedPropertyRepository : EfRepository<LocalizedProperty, int>, ILocalizedPropertyRepository
+    public class LocalizedPropertyRepository : EfRepository<LocalizedProperty, ulong>, ILocalizedPropertyRepository
     {
         public LocalizedPropertyRepository(DbContext context) : base(context)
         {
